chore: Moved service/kernel/execution-context.registry.mts to service/execution-context.registry.mts.
