+++ packages/architectura/src/core/rich-server-response.mts
@@ -2,9 +2,10 @@ import { ServerResponse as HTTPServerResponse } from "node:http";
 
 import { type Gzip, createGzip } from "node:zlib";
 
+import { KernelService } from "../service/execution-context/execution-context.registry.mjs";
+
 import { ExecutionContextService } from "../service/execution-context/execution-context.service.mjs";
 
-import { KernelService } from "../service/kernel/execution-context.registry.mjs";
 
 import type { RichClientRequest } from "./rich-client-request.mjs";
