+++ packages/architectura/src/service/execution-context/_index.mts
@@ -1,1 +1,2 @@
+export * from './execution-context.registry.mjs';
 export * from './execution-context.service.mjs';
