--- packages/architectura/src/service/kernel/_index.mts
@@ -1,1 +1,0 @@
-export * from './execution-context.registry.mjs';
